// web/src/types/chat/conversation_message.ts
export interface ConversationMessageBase extends SeqMessageBase {
	/** 应用消息 ID */
	app_message_id: string
	/** 发送者 ID */
	sender_id: string
	/** 发送时间 */
	send_time: number
	/** 消息状态 */
	status: ConversationMessageStatus
	/** 未读消息数 */
	unread_count: number
	/** 话题 ID */
	topic_id?: string
	/** 是否撤回 */
	revoked?: boolean
	/** 本地临时数据存储 */
	temp_custom_data?: Record<string, unknown>
}
// ...
export interface GroupAddMemberMessage extends ConversationMessageBase {
	type: ControlEventMessageType.GroupAddMember
	/** 未读数 */
	unread_count: number
	/** 发送时间 */
	send_time: number
	/** 状态 */
	status: ConversationMessageStatus
	/** 群新增成员 */
	group_users_add: {
		/** 操作人ID */
		operate_user_id: string
		/** 群ID */
		group_id: string
		/** 用户ID列表 */
		user_ids: string[]
		/** 会话ID */
		conversation_id: string
	}
}
// ...
export const enum ConversationMessageStatus {
	/** 未读 */
	Unread = "unread",
	/** 已读 */
	Read = "read",
	/** 已看 */
	Seen = "seen",
	/** 已撤回 */
	Revoked = "revoked",
}